integrated database and created API

+++ LearnManagerAPI/DTOs/QuizSubmitDto.cs
@@ -2,6 +2,6 @@ namespace LearnManagerAPI.DTOs;
 
 public class QuizSubmitDto
 {
-    public int QuizId { get; set; }
-    public Dictionary<int, string> Answers { get; set; }
+    public long QuizId { get; set; }
+    public Dictionary<long, string> Answers { get; set; }
 }
