+++ LearnManagerAPI/Services/IUserService.cs
@@ -5,10 +5,10 @@ namespace LearnManagerAPI.Services
     public interface IUserService
     {
         List<User> GetAllUsers();
-        User GetUserById(int id);
+        User GetUserById(long id);
         User GetUserByEmail(string email);
         User Createuser(User user);
-        User UpdateUser(int id, User user);
-        void DeleteUser(int id);
+        User UpdateUser(long id, User user);
+        void DeleteUser(long id);
     }
 }
