+++ LearnManagerAPI/Services/Interfaces/IEnrollmentService.cs
@@ -4,7 +4,7 @@ namespace LearnManagerAPI.Services.Interfaces
 {
     public interface IEnrollmentService
     {
-        Task<Enrollment> EnrollAsync(int studentId, int courseId);
-        Task<IEnumerable<Enrollment>> GetEnrollmentsByStudentAsync(int studentId);
+        Task<Enrollment> EnrollAsync(long studentId, long courseId);
+        Task<IEnumerable<Enrollment>> GetEnrollmentsByStudentAsync(long studentId);
     }
 }
