+++ LearnManagerAPI/Services/Interfaces/IQuizService.cs
@@ -5,8 +5,8 @@ namespace LearnManagerAPI.Services.Interfaces
 {
     public interface IQuizService
     {
-        Task<Quiz> CreateQuizAsync(int courseId, Quiz quiz);
-        Task<Quiz> GetQuizByIdAsync(int id);
-        Task<int> SubmitQuizAsync(QuizSubmitDto dto, int studentId);
+        Task<Quiz> CreateQuizAsync(long courseId, Quiz quiz);
+        Task<Quiz> GetQuizByIdAsync(long id);
+        Task<int> SubmitQuizAsync(QuizSubmitDto dto, long studentId);
     }
 }
